Added progress when scanning files and fixed some bugs in the API
