+++ lightrag/api/static/index.html
@@ -98,7 +98,7 @@
         </div>
     </div>
 
-    <script src="/js/lightrag_api.js"></script>
+    <script src="/js/api.js"></script>
 
 </body>
 </html>
